Formatted output
Though the addresses of the edges changes because of copying, the
edge.oid remains unchanged. So, edge.oid is used as the key for
formatting the output. There is only one edge present in INVDS_2.keys()
but not in score.keys() because the score is not calculated for the
final edge of the postorder traversal as it redundant.